requirements.txt
streamlit
ultralytics
opencv-python
opencv-python-headless
Pillow
numpy
onnx
fastapi
python-dotenv

# for linting and formatting
pylint
isort
black 